refactor: [C4R-1209] renamed the dto to interface
