+++ backend/src/modules/organization/interfaces/OrganizationWithPracticePrograms.interface.ts
@@ -1,6 +1,6 @@
 import { PracticeProgram } from 'src/modules/practice-program/entities/practice-program.entity';
 import { Domain } from 'src/shared/entities';
-import { ContactPerson } from '../dto/contact-person.dto';
+import { ContactPerson } from './contact-person.interface';
 
 export interface OrganizationWithPracticePrograms {
   id: number;
